remove undefined props from getServerSideProps to be able to JSON.stringify them
