+++ catalogue/webapp/pages/images.tsx
@@ -1,5 +1,5 @@
 import { GetServerSideProps, NextPage } from 'next';
-import { useEffect, useState, useContext, ReactElement } from 'react';
+import { useEffect, useState, ReactElement } from 'react';
 import Router from 'next/router';
 import Head from 'next/head';
 import {
@@ -28,13 +28,13 @@ import SearchForm from '@weco/common/views/components/SearchForm/SearchForm';
 import { getImages } from '../services/catalogue/images';
 import useSavedSearchState from '@weco/common/hooks/useSavedSearchState';
 import useHotjar from '@weco/common/hooks/useHotjar';
-import TogglesContext from '@weco/common/views/components/TogglesContext/TogglesContext';
 import SearchTabs from '@weco/common/views/components/SearchTabs/SearchTabs';
 import SearchNoResults from '../components/SearchNoResults/SearchNoResults';
 import {
   getGlobalContextData,
   WithGlobalContextData,
 } from '@weco/common/views/components/GlobalContextProvider/GlobalContextProvider';
+import { removeUndefinedProps } from '@weco/common/utils/json';
 
 type Props = {
   results?: CatalogueResultsList<Image> | CatalogueApiError;
@@ -284,12 +284,12 @@ export const getServerSideProps: GetServerSideProps<Props> = async context => {
     : undefined;
 
   return {
-    props: {
+    props: removeUndefinedProps({
       results: imagesOrError,
       imagesRouteProps: params,
       apiProps,
       globalContextData,
-    },
+    }),
   };
 };
