--- common/test/utils/json.test.js
@@ -1,18 +1,0 @@
-// @flow
-import { removeEmptyProps } from '../../utils/json';
-
-describe('remove empty', () => {
-  it('should remove null and undefined values', () => {
-    const obj = {
-      thing: 'that exists',
-      anotherThatIsNull: null,
-      iAmUndefined: undefined,
-    };
-
-    const emptiedObject = removeEmptyProps(obj);
-
-    expect(emptiedObject.thing).toBe('that exists');
-    expect(emptiedObject.anotherThatIsNull).toBeUndefined();
-    expect(emptiedObject.iAmUndefined).toBeUndefined();
-  });
-});
