--- common/utils/json.js
@@ -1,8 +1,0 @@
-// @flow
-export function removeEmptyProps(obj: Object): Object {
-  return JSON.parse(
-    JSON.stringify(obj, function(key, value) {
-      return value === null || value === undefined ? undefined : value;
-    })
-  );
-}
